Fix sonar issues
