+++ src/main/java/com/checkmarx/flow/dto/HookType.java
@@ -1,0 +1,5 @@
+package com.checkmarx.flow.dto;
+
+public enum HookType {
+    PUSH, PULL_REQUEST
+}
