+++ src/test/java/com/checkmarx/flow/cucumber/integration/end2end/genericendtoend/GenericEndToEndSteps.java
@@ -7,6 +7,7 @@ import com.checkmarx.flow.config.GitHubProperties;
 import com.checkmarx.flow.config.JiraProperties;
 import com.checkmarx.flow.cucumber.common.utils.TestUtils;
 
+import com.checkmarx.flow.dto.HookType;
 import io.cucumber.java.After;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
