--- src/test/java/com/checkmarx/flow/cucumber/integration/end2end/genericendtoend/HookType.java
@@ -1,5 +1,0 @@
-package com.checkmarx.flow.cucumber.integration.end2end.genericendtoend;
-
-enum HookType {
-    PUSH, PULL_REQUEST
-}
